[#4] refactor: 패키지 자료 구조 변경
- User 패키지 자료 구조 변경
