+++ src/main/java/com/sparta/tentrillion/board/controller/BoardController.java
@@ -7,8 +7,8 @@ import com.sparta.tentrillion.board.dto.response.BoardResponseDto;
 import com.sparta.tentrillion.board.service.BoardService;
 import com.sparta.tentrillion.global.argumentResolver.annotation.LoginUser;
 import com.sparta.tentrillion.global.exception.BusinessException;
-import com.sparta.tentrillion.user.User;
-import com.sparta.tentrillion.user.UserService;
+import com.sparta.tentrillion.user.entity.User;
+import com.sparta.tentrillion.user.service.UserService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
