+++ src/main/java/com/sparta/tentrillion/board/entity/InvitedBoard.java
@@ -1,7 +1,7 @@
 package com.sparta.tentrillion.board.entity;
 
 import com.sparta.tentrillion.global.TimeStamp;
-import com.sparta.tentrillion.user.User;
+import com.sparta.tentrillion.user.entity.User;
 import jakarta.persistence.*;
 import lombok.*;
