+++ src/main/java/com/sparta/tentrillion/board/repository/BoardQRepository.java
@@ -1,7 +1,7 @@
 package com.sparta.tentrillion.board.repository;
 
 import com.sparta.tentrillion.board.entity.Board;
-import com.sparta.tentrillion.user.User;
+import com.sparta.tentrillion.user.entity.User;
 
 import java.util.List;
