+++ src/main/java/com/sparta/tentrillion/board/repository/InvitedBoardRepository.java
@@ -2,7 +2,7 @@ package com.sparta.tentrillion.board.repository;
 
 import com.sparta.tentrillion.board.entity.Board;
 import com.sparta.tentrillion.board.entity.InvitedBoard;
-import com.sparta.tentrillion.user.User;
+import com.sparta.tentrillion.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
