+++ src/main/java/com/sparta/tentrillion/board/service/BoardService.java
@@ -8,8 +8,8 @@ import com.sparta.tentrillion.board.entity.InvitedBoard;
 import com.sparta.tentrillion.board.repository.BoardRepository;
 import com.sparta.tentrillion.board.repository.InvitedBoardRepository;
 import com.sparta.tentrillion.global.exception.BusinessException;
-import com.sparta.tentrillion.user.User;
-import com.sparta.tentrillion.user.UserService;
+import com.sparta.tentrillion.user.entity.User;
+import com.sparta.tentrillion.user.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
