+++ src/main/java/com/sparta/tentrillion/card/Card.java
@@ -3,7 +3,7 @@ package com.sparta.tentrillion.card;
 import com.sparta.tentrillion.comment.entity.Comment;
 import com.sparta.tentrillion.global.TimeStamp;
 import com.sparta.tentrillion.stat.entity.Stat;
-import com.sparta.tentrillion.user.User;
+import com.sparta.tentrillion.user.entity.User;
 import jakarta.persistence.*;
 import lombok.*;
