+++ src/main/java/com/sparta/tentrillion/comment/entity/Comment.java
@@ -2,7 +2,7 @@ package com.sparta.tentrillion.comment.entity;
 
 import com.sparta.tentrillion.card.Card;
 import com.sparta.tentrillion.global.TimeStamp;
-import com.sparta.tentrillion.user.User;
+import com.sparta.tentrillion.user.entity.User;
 import jakarta.persistence.*;
 import lombok.*;
