+++ src/main/java/com/sparta/tentrillion/mail/MailService.java
@@ -4,8 +4,8 @@ import com.sparta.tentrillion.global.exception.BusinessException;
 import com.sparta.tentrillion.global.exception.ErrorCode;
 import com.sparta.tentrillion.mail.dto.MailRequestDto;
 import com.sparta.tentrillion.mail.dto.VerifyRequestDto;
-import com.sparta.tentrillion.user.User;
-import com.sparta.tentrillion.user.UserRepository;
+import com.sparta.tentrillion.user.entity.User;
+import com.sparta.tentrillion.user.repository.UserRepository;
 import jakarta.mail.MessagingException;
 import jakarta.mail.internet.MimeMessage;
 import lombok.RequiredArgsConstructor;
